+++ src/model/variable.ts
@@ -32,7 +32,12 @@ export interface Variable {
     dtype: string;
     units: string;
     title: string;
-    tileSourceOptions: TileSourceOptions;
+    // tileUrl is new since xcube 0.11
+    tileUrl?: string;
+    tileLevelMin?: number;
+    tileLevelMax?: number;
+    // tileSourceOptions are longer used since xcube 0.11
+    tileSourceOptions?: TileSourceOptions;
     colorBarName: string;
     colorBarMin: number;
     colorBarMax: number;